+++ src/assertions/xml.ts
@@ -1,4 +1,7 @@
 import { XMLParser } from 'fast-xml-parser';
+import type { GradingResult } from '../types';
+import type { Assertion } from '../types';
+import type { AssertionValue } from '../types';
 
 export function validateXml(
   xmlString: string,
@@ -63,3 +66,36 @@ export function containsXml(
 
   return { isValid: false, reason: 'No valid XML content found matching the requirements' };
 }
+
+export const handleIsXml = (
+  assertion: Assertion,
+  baseType: string,
+  outputString: string,
+  renderedValue: AssertionValue | undefined,
+  inverse: boolean,
+): GradingResult => {
+  let requiredElements: string[] | undefined;
+  if (typeof renderedValue === 'string') {
+    requiredElements = renderedValue.split(',').map((el) => el.trim());
+  } else if (Array.isArray(renderedValue) && renderedValue.length > 0) {
+    requiredElements = renderedValue.map((el) => el.toString());
+  } else if (typeof renderedValue === 'object' && Object.keys(renderedValue).length > 0) {
+    if ('requiredElements' in renderedValue && Array.isArray(renderedValue.requiredElements)) {
+      requiredElements = renderedValue.requiredElements.map((el) => el.toString());
+    } else {
+      throw new Error('xml assertion must contain a string, array value, or no value');
+    }
+  }
+
+  const result = (baseType === 'is-xml' ? validateXml : containsXml)(
+    outputString,
+    requiredElements,
+  );
+  const pass = result.isValid !== inverse;
+  return {
+    pass,
+    score: pass ? 1 : 0,
+    reason: pass ? 'Assertion passed' : result.reason,
+    assertion,
+  };
+};
